requirements.dev.txt
tox
pytest
flake8>=3.9.2,<3.10
pytest-django
pytest-watch
pytest-mock